Bill S33 version original - Patrick M. Gallivan
BILL NUMBER: S33 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the penal law, in relation to establishing the crime of making a terroristic threat against a police officer PURPOSE: This bill would establish the crime of making a terroristic threat against a police officer. SUMMARY OF PROVISIONS: Section 1 amends the penal law to create a new section establishing the crime of making a terroristic threat against a police officer.