Bill S79A version original - Patrick M. Gallivan
BILL NUMBER: S79 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to establishing the crime of doxing a police officer or a peace officer PURPOSE: This legislation establishes the criminal offense of doxing a police or peace officer. Makes such offense a class D felony.