Bill S64 version original - Pete Harckham
BILL NUMBER: S64 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the real property tax law, in relation to taxation of state owned land PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to provide equity in relation to the taxation of state owned lands. SUMMARY OF PROVISIONS: Section 1 amends section 532 of the real property tax law by adding section 2.