Bill S66A version original - Pete Harckham
BILL NUMBER: S66 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the education law, in relation to granting licensed psychologists prescriptive authority PURPOSE: To allow certain licensed psychologists to prescribe medication. SUMMARY OF PROVISIONS: Section 1 amends section 7601-a of the education law to include prescription of appropriate drugs as within the definition of the prac- tice of psychology. Section 2 adds a new section 7601-b to the education law.