Bill S37 version original - Patrick M. Gallivan
BILL NUMBER: S37 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the labor law, in relation to clarifying methods for the payment of wages and authorizing the payment of wages by use of payroll cards PURPOSE: The bill would allow New York employers to provide for the payment of wages on payroll cards. SUMMARY OF PROVISIONS: Section 1 amends the heading of section 192 of the Labor Law.