Bill S48 version original - James Tedisco
BILL NUMBER: S48 SPONSOR: TEDISCO TITLE OF BILL: CONCURRENT RESOLUTION OF THE SENATE AND ASSEMBLY proposing an amendment to section 14 of article 3 of the constitution, in relation to prohibit- ing the passage of bills between the hours of midnight and eight o'clock a.m. and requiring two-thirds approval for messages of necessity PURPOSE OR GENERAL IDEA OF BILL: Prohibits the passage of bills by the legislature between the hours of midnight and eight o'clock a.m.