Bill S77 version original - Patrick M. Gallivan
BILL NUMBER: S77 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to establishing the crime of failure to retreat PURPOSE: This legislation establishes the criminal offense of failure to halt or to retreat, with exceptions, when instructed to do so by a police or peace officer. Makes such offense a class D felony.