Bill S65 version original - Pete Harckham
BILL NUMBER: S65 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to distinctive license plates for society for the prevention of cruelty to animals officers PURPOSE: To establish distinctive license plates for society for the prevention of cruelty to animals officers. SUMMARY OF PROVISIONS: Section 1 sets the provisions by which a Society for the Prevention of Cruelty to Animals officer or their spouse may purchase a distinctive license plate.