Bill S57 version original - Pete Harckham
BILL NUMBER: S57 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the tax law, in relation to enacting the "STAR credit bill of rights" and establishing the office of STAR ombudsman PURPOSE: Enacts the "STAR credit bill of rights and establishes the office of STAR ombudsman to ensure STAR recipients are provided a better under- standing of their rebate and their inquiries receive prompt responses. SUMMARY OF PROVISIONS: Section 1 establishes the effective date.