Bill S39 version original - Patrick M. Gallivan
BILL NUMBER: S39 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to police staffing at certain parks PURPOSE: The purpose of this bill is to ensure that the Commissioner of Parks, Recreation and Historic Preservation provide for the health, safety and welfare of the public using the facilities under its jurisdiction by ensuring adequate state park police officers at such facilities. SUMMARY OF PROVISIONS: Section 1.