Bill S42 version original - Patrick M. Gallivan
BILL NUMBER: S42 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the executive law, in relation to certain reimbursement rates PURPOSE: To reinstate the state funded reimbursement for housing parole violators in local correctional facilities eliminated in 2009. SUMMARY OF PROVISIONS: Section 1 of the bill amends the executive law, section 259-i by adding a new subdivision 2-a. Section 2 of the bill provides the effective date.