Bill S47 version original - James Tedisco
BILL NUMBER: S47 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the legislative law, in relation to bills with a thresh- old level of co-sponsorship PURPOSE: To allow for a piece of legislation that has garnered the co-sponsorship of 76 members of the Assembly and 32 members in the Senate to be brought to the floor in that respective house for a vote. SUMMARY OF PROVISIONS: Section 1: Amends the legislative law by adding a new section 33-a to read as follows: 33-a.