Bill S20A version original - Andrew Gounardes
BILL NUMBER: S20 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the public authorities law, in relation to membership on the board of the metropolitan transportation authority PURPOSE OR GENERAL IDEA OF BILL: This legislation will add four rider representatives to the MTA Board as voting members. SUMMARY OF PROVISIONS: Sections one and two of this bill amend Public Authorities Law section 1263(a)(1) to add four new voting members and three non-voting members to the MTA Board.