Bill S78 version original - Patrick M. Gallivan
BILL NUMBER: S78 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the tax law, in relation to excise taxes on cigars and premium cigars PURPOSE: To establish the rate of the tobacco products tax exclusively for premi- um cigars or cigars at seventy-five percent of the wholesale price or fifty cents, whichever is less. SUMMARY OF PROVISIONS: Section 1 establishes a definition of "premium cigar" under section 470 of the tax law.