Bill S16 version original - Andrew Gounardes
BILL NUMBER: S16 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the real property law, in relation to sale or rent of illegally converted dwellings PURPOSE OR GENERAL IDEA OF BILL: This bill prohibits the listing for sale, sale and offering to sell, listing for rent, renting and offering to rent any illegally converted dwelling. The bill establishes a license revocation for real estate brokers, agents and salespeople who violate the provisions of this act.