Bill S44 version original - James Tedisco
BILL NUMBER: S44 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the education law, in relation to enacting "Jacobe's law" PURPOSE: To ensure a "reasonable and good faith effort" is made by school admin- istrators to notify parents or guardians when there is a verified inci- dent of harassment, bullying or discrimination in school. SUMMARY OF PROVISIONS: Section 1 of the bill cites the act as Jacobe's Law.