Bill S18 version original - Andrew Gounardes
BILL NUMBER: S18 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the insurance law, in relation to requiring business interruption insurance to cover disease outbreaks PURPOSE OR GENERAL IDEA OF BILL: This bill would require business interruption insurance offered to small businesses and/or not-for profits of 250 or less employees to cover disease outbreaks. SUMMARY OF PROVISIONS: Section 1 amends the insurance law by adding a new section 346.