Bill S97 version original - James Tedisco
BILL NUMBER: S97 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the state finance law, in relation to enacting the "over-expenditure, under-expenditure, transfer notification (OUT) act" PURPOSE: Requires that the Comptroller be immediately notified of any over-expen- diture of taxpayer monies and that all leftover funds are allocated to the general fund to pay down debt. SUMMARY OF PROVISIONS: Section 1: Cites the act as the "over-expenditure, under-expenditure, transfer-notification (OUT) act.