Bill S142 version original - Michael Gianaris
BILL NUMBER: S142 SPONSOR: GIANARIS TITLE OF BILL: An act to amend the agriculture and markets law, in relation to restricting the performance of surgical devocalization procedures on dogs and cats PURPOSE: To restrict the practice of devocalizing dogs and cats to medical neces- sity. SUMMARY OF PROVISIONS: Section one amends the agriculture and markets law by adding a new section 365-a restricting the practice of devocalizing a dog or cat.