Bill S109 version original - Liz Krueger
BILL NUMBER: S109 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the legislative law, in relation to requiring the inclu- sion of certain information in fiscal notes in retirement bills and on bills affecting political subdivisions PURPOSE: Creates a statutory requirement for legislators to include fiscal notes in bills which would impact the expenses and revenues of political subdivisions by at least $250,000. SUMMARY OF PROVISIONS: Section 1. Section 50 of the Legislative Law is amended. 50.