Bill S126 version original - Sean Ryan
BILL NUMBER: S126 SPONSOR: RYAN TITLE OF BILL: An act to amend the executive law, in relation to the education and experience of members of the state board of parole PURPOSE: The purpose of this bill is to expand the eligible pool of candidates for the parole board. SUMMARY OF PROVISIONS: Section 1 amends subdivision 2 of section 259-b of the executive law to permit five years of relevant experience to serve as a substitution for a bachelor's degree. Section 2 is the effective date.