Bill S140 version original - Sean Ryan
BILL NUMBER: S140 SPONSOR: RYAN TITLE OF BILL: An act to amend the penal law, in relation to revoking a license to carry, possess, repair and dispose of firearms for a conviction for failure to safely store rifles, shotguns, and firearms PURPOSE: To ensure that persons engaging in unsafe storage of firearms cannot possess firearms. SUMMARY OF PROVISIONS: Section 1 amends paragraph (a) of subdivision 11 of section 400.00 of the 2 penal law to add convictions of a licensee under section 265.