Bill S95 version original - James Tedisco
BILL NUMBER: S95 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the penal law and the highway law, in relation to violence committed on school grounds PURPOSE OR GENERAL IDEA OF BILL: This bill would increase the penalties for assault and abduction that take place on school grounds, including nursery schools, and college campuses. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Names this legislation "Suzanne's Law." Section 2: Adds a new Section 70.