Bill S96 version original - James Tedisco
BILL NUMBER: S96 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the agriculture and markets law, in relation to expand- ing the definition of aggravated cruelty to animals to include harm to animals during the commission of a felony PURPOSE: To enhance laws relating to animal cruelty by including harm caused during the commission of a felony or flight there from. SUMMARY OF PROVISIONS: This act shall be known and may be cited as "Kirby & Quigley's Law.