Bill S82 version original - Patrick M. Gallivan
BILL NUMBER: S82 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to creating the crime of aggravated offering of a false accusa- tion against a police officer PURPOSE: This bill establishes the crime of aggravated offering of a false accu- sation against a police or peace officer when, knowing the information reported to be false or baseless, a person files such a report. Makes the offense a class D felony.