Bill S102 version original - James Tedisco
BILL NUMBER: S102 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the executive law, in relation to grants to assist local governments with reducing paper PURPOSE OR GENERAL IDEA OF BILL: This bill provides incentives via grant rewards to reduce paperwork at the municipal and local levels of government and implement digital alternatives. SUMMARY OF SPECIFIC PROVISIONS: Article 7 of the executive law is amended by adding a new section 171 to read as follows: § 171.