Bill S107 version original - James Tedisco
BILL NUMBER: S107 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the tax law, in relation to establishing a tax credit for the cost of spay or neuter services PURPOSE: To grant an individual taxpayer a credit of eighty percent of the actual cost of having their cats or dogs spayed or neutered. Credit shall not exceed two hundred dollars.