Bill S125 version original - Patrick M. Gallivan
BILL NUMBER: S125 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the general construction law, in relation to allowing for online publication of public notices PURPOSE OR GENERAL IDEA OF BILL: Allows for online publication of public notices. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends the General Construction law to permit a public notice, required by law to be published or advertised in a newspaper, to instead be published in an online news publication.