Bill S87 version original - Patrick M. Gallivan
BILL NUMBER: S87 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the financial services law, the tax law and the public health law, in relation to enacting the Health Care Tax Reform Act PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to reduce the amount of taxes imposed on health care and health insurance consumers. These taxes collectively are over $5 Billion annually, which exceeds the amount of annual fran- chise taxes paid by corporations.