Bill S167 version original - Michael Gianaris
BILL NUMBER: S0167 SPONSOR: GIANARIS TITLE OF BILL: An act to amend the criminal procedure law and the penal law, in relation to abolishing citizen's arrests; and to repeal certain provisions of the criminal procedure law and the family court act relat- ing thereto PURPOSE: To repeal citizen's arrest laws in the state of New York. SUMMARY OF PROVISIONS: Section one amends section 140.45 of the criminal procedure law by removing reference to section 140.40 of the criminal procedure law.