Bill S160B version original - Brad Hoylman-Sigal
BILL NUMBER: S160 SPONSOR: HOYLMAN TITLE OF BILL: An act to amend the judiciary law, in relation to audio-visual coverage of judicial proceedings; and to repeal section 218 of the judiciary law and section 52 of the civil rights law relating thereto PURPOSE: Provides for audio-visual recordings in judicial proceedings; repealer. SUMMARY OF PROVISIONS: Section 1 repeals Section 218 of the Judiciary law and adds a new Section 218 that provides for audio-visual recordings in judicial proceedings.