Bill S189 version original - Patrick M. Gallivan
BILL NUMBER: S189 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the executive law, in relation to third party statements to the parole board PURPOSE OF THE BILL: To amend section 259-i of the Executive Law regarding the submission of written statements by the public. SUMMARY OF PROVISIONS: This bill amends Executive Law § 259-1(2)(c)(A) to explicitly codify the public's right to submit written statements regarding release determi- nations.