Bill S180 version original - Robert Jackson
BILL NUMBER: S180 SPONSOR: JACKSON TITLE OF BILL: An act to amend the real property law, in relation to establishing secu- rity deposit options for certain tenants PURPOSE: The purpose of this bill is to provide financial alternatives to pay rental security deposits for tenants of market and rent regulated units. SUMMARY OF PROVISIONS: Section 1 adds a new Section 238-B to the Real Property Law.