Bill S197 version original - Patrick M. Gallivan
BILL NUMBER: S197 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the education law, in relation to the establishment of an extraordinary needs reserve fund PURPOSE OR GENERAL IDEA OF BILL: This legislation affords school districts the ability to establish an extraordinary needs reserve fund for the expenses of providing special education. SUMMARY OF SPECIFIC PROVISIONS: Section 1.