Bill S218 version original - Zellnor Myrie
BILL NUMBER: S218 SPONSOR: MYRIE TITLE OF BILL: An act to amend the executive law, in relation to duties of notaries with respect to instruments conveying residential real property PURPOSE: The purpose of the bill is to establish a clear set of notary duties in an effort to prevent deed theft. SUMMARY OF PROVISIONS: Section 1 amends the Executive Law with the addition of section 137-b, which outlines the duties of notaries with respect to instruments conveying residential real property.