Bill S221A version original - Zellnor Myrie
BILL NUMBER: S221 SPONSOR: MYRIE TITLE OF BILL: An act to amend the real property actions and proceedings law, in relation to establishing the tenant opportunity to purchase act SUMMARY OF PROVISIONS: Section 1 provides the title of the bill. Section 2 provides legislative findings regarding the significant unmet need of affordable housing for low-income people and families.