Bill S192 version original - Patrick M. Gallivan
BILL NUMBER: S192 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the tax law, in relation to providing a tax credit for a portion of the cost of qualified densified biomass fuel expenditures; and providing for the repeal of such provisions upon expiration thereof PURPOSE: This bill creates a tax credit for the purchase of biomass fuel for residential use, typically wood pellets for use in a wood pellet boiler. SUMMARY OF PROVISIONS: Section 1.