Bill S170 version original - Michael Gianaris
BILL NUMBER: S170 SPONSOR: GIANARIS TITLE OF BILL: An act to amend the election law, in relation to allowing for polling place voter registration for any qualified person who is not registered to vote PURPOSE: Provides that qualified persons who are not registered to vote may register on a day of election at their polling places. SUMMARY OF PROVISIONS: Section 1 amends 5-210 of election law by adding a new subdivision 1-a which enables qualified persons to register to vote.