Bill S185 version original - Patrick M. Gallivan
BILL NUMBER: S185 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to leaving the scene of an incident without reporting PURPOSE: To increase the penalty for leaving the scene of an incident without reporting when there is serious physical injury or death. SUMMARY OF PROVISIONS: Section 1 amends paragraph c of subdivision 2 of section 600 of the vehicle and traffic law. Section 2 provides the effective date.