Bill S169 version original - Michael Gianaris
BILL NUMBER: S169 SPONSOR: GIANARIS TITLE OF BILL: An act to amend the labor law, in relation to designating discrimination against employees for failure to meet certain ticket quotas as a class A misdemeanor PURPOSE: To improve the transparency and reliability of the public contracting process by municipalities. SUMMARY OF PROVISIONS: Section one amends section 215-a of the Labor Law, by adding a new subdivision three.