Bill S251 version original - Patrick M. Gallivan
BILL NUMBER: S251 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the criminal procedure law, in relation to examination of witnesses PURPOSE: To add elderly witnesses, age 75 years or older, to the category of witnesses who may be eligible to be examined conditionally, pursuant to Article 660 of the Criminal Procedure Law, in order that such testimony may be received into evidence at subsequent proceedings in or related to the action. SUMMARY OF PROVISIONS: Section 1 amends § 660.