Bill S293 version original - James Tedisco
BILL NUMBER: S293 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the agriculture and markets law, in relation to imposing additional penalties for animal abuse PURPOSE OR GENERAL IDEA OF BILL: To prevent the crime of animal cruelty through evaluation and treatment of offenders. SUMMARY OF SPECIFIC PROVISIONS: Provides that a person convicted of animal abuse or neglect would be required to undergo a psychiatric evaluation and treatment as warranted.