Bill S301 version original - Julia Salazar
BILL NUMBER: S301 SPONSOR: SALAZAR TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to second felony drug offenders; and to repeal subdivision 4 of section 70.70 of the penal law relating thereto PURPOSE: To ensure that all individuals facing drug charges are treated similar- ly, without regard to prior non-drug related convictions. SUMMARY OF PROVISIONS: Section 1: Amends paragraph (a) of subdivision 3 of section 70.70 of the penal law.