Bill S300 version original - Andrea Stewart-Cousins
BILL NUMBER: S300 SPONSOR: STEWART-COUSINS TITLE OF BILL: An act to amend the education law, in relation to establishing the community schools act; and making an appropriation therefor PURPOSE: The purpose of this bill is to ensure high-quality educational opportu- nities, and improve educational outcomes for all students through investment in sustainable community schools. SUMMARY OF PROVISIONS: Section one names the bill The Community Schools Act.