Bill S252 version original - José M. Serrano
BILL NUMBER: S252 SPONSOR: SERRANO TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to the reporting of alienation of municipal parkland PURPOSE OR GENERAL IDEA OF BILL: Municipally owned parkland is vital to local communities. Once lost, municipal parkland is difficult to recover.