Bill S281 version original - Andrew Gounardes
BILL NUMBER: S281 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the criminal procedure law, in relation to periods of limitation in calculating the timeliness of prosecution PURPOSE OR GENERAL IDEA OF BILL: To exclude any period of time during which an individual was unjustly convicted or imprisoned when calculating the time limitation applicable to commencement of a criminal action. SUMMARY OF PROVISIONS: Section 1 amends subdivision 4 of section 30.