Bill S264A version original - Andrea Stewart-Cousins
BILL NUMBER: S264 SPONSOR: STEWART-COUSINS TITLE OF BILL: An act to amend the executive law, in relation to establishing and main- taining an emergency evacuation plan for individuals with disabilities PURPOSE: The purpose of this bill is to ensure that there is an efficient emer- gency evacuation plan for people with disabilities who work or live in high-rise buildings in New York State. SUMMARY OF PROVISIONS: Section one amends the Executive Law by adding two new sections 378-a and 378-b.