Bill S285 version original - Rachel May
BILL NUMBER: S285 SPONSOR: MAY TITLE OF BILL: An act to amend the education law, in relation to adding arts and music education into the curriculum for public school students PURPOSE: Ensure arts and music education are incorporated into the public school curriculum for all students. SUMMARY OF PROVISIONS: Section 1 sets for the legislative intent that discusses the importance of arts and music education.