Bill S282 version original - Andrew Gounardes
BILL NUMBER: S282 SPONSOR: GOUNARDES TITLE OF BILL: An act to repeal section 270.25 of the criminal procedure law relating to peremptory challenges of jurors PURPOSE OR GENERAL IDEA OF BILL: Relates to abolishing peremptory challenges of jurors in criminal cases. SUMMARY OF PROVISIONS: Section 1 repeals section 270.25 of the criminal procedure law. Section 2 states the effective date. JUSTIFICATION: In 1986, the Supreme Court of the United States found in Batson v.