Bill S263 version original - Andrea Stewart-Cousins
BILL NUMBER: S263 SPONSOR: STEWART-COUSINS TITLE OF BILL: An act to amend the election law, in relation to prohibiting the suppression of voters PURPOSE: The purpose of this act is to create a new electoral crime of voter suppression, punishable as a misdemeanor. SUMMARY OF PROVISIONS: Section one names the bill the "deceptive practices and voter suppression prevention act.