Bill S376 version original - Kevin S. Parker
BILL NUMBER: S376 SPONSOR: PARKER TITLE OF BILL: An act to amend the transportation corporations law and the public service law, in relation to authorizing broadband franchises in cities with a population of one million or more PURPOSE OR GENERAL IDEA OF BILL: To increase competition between broadband providers through the develop- ment of a franchise system. SUMMARY OF SPECIFIC PROVISIONS: Section 1.