Bill S386 version original - George Borrello
BILL NUMBER: S386 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the election law, in relation to enacting the "voter registration integrity act" PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to ensure that newly registered voters are citizens of the United States. SUMMARY OF SPECIFIC PROVISIONS: Section 1 names the bill the "voter registration integrity act.