Bill S383 version original - Cordell Cleare
BILL NUMBER: S383 SPONSOR: CLEARE TITLE OF BILL: An act to amend the correction law, in relation to requiring gynecologi- cal services to be offered to certain incarcerated individuals PURPOSE OR GENERAL IDEA OF BILL: Currently, there is no state law that requires routine cervical screen- ings for incarcerated individuals. Regular gynecological visits are a standard and necessary part of general health care for anyone with a cervix.