Bill S387 version original - George Borrello
BILL NUMBER: S387 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the public lands law, in relation to the payment of state aid on account of certain state lands in the town of New Hudson, in the county of Allegany PURPOSE: Provides for the payment of State aid to the Town of New Hudson, in the County of Allegany, and special districts therein for State land in such town operated by the Department of Environmental Conservation (DEC). SUMMARY OF PROVISIONS: Section 1.