Bill S381 version original - Cordell Cleare
BILL NUMBER: S381 SPONSOR: CLEARE TITLE OF BILL: An act to amend the executive law, in relation to providing a right to representation by counsel for certain actions of the parole board PURPOSE OR GENERAL IDEA OF BILL: To provide a right to counsel to individuals under consideration for parole. SUMMARY OF SPECIFIC PROVISIONS: Provides a right to counsel to individuals under consideration for parole.