Bill S338 version original - Robert Jackson
BILL NUMBER: S338 SPONSOR: JACKSON TITLE OF BILL: An act to amend the general business law, in relation to discounted admission and memberships for certain individuals PURPOSE: The purpose of this legislation is to provide admission and membership discounts to individuals with disabilities and cultural and enter- tainment institutions. SUMMARY OF PROVISIONS: Section 1 amends the General Business Law by adding a new Section 396-cccc.