Bill S400 version original - Andrea Stewart-Cousins
BILL NUMBER: S400 SPONSOR: STEWART-COUSINS TITLE OF BILL: An act to amend the local finance law, in relation to bonds and notes of the city of Yonkers PURPOSE: Extender law - this bill extends the marketing limitation on serial bonds/notes of Yonkers from June 30, 2023 to June 30, 2024. SUMMARY OF SPECIFIC PROVISIONS: Section one extends the City of Yonkers' authorization to issue bonds and notes through negotiated sale to June 30, 2024. Section two is the effective date.