Bill S432 version original - George Borrello
BILL NUMBER: S432 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the tax law, in relation to establishing a first perma- nent payroll employee tax credit PURPOSE: The purpose of this bill is to establish a first permanent payroll employee tax credit. SUMMARY OF PROVISIONS: Section 1 of the bill amends Section 210-B of the Tax Law by adding a new subdivision 55.