Bill S394 version original - Cordell Cleare
BILL NUMBER: S394 SPONSOR: CLEARE TITLE OF BILL: An act to amend the education law, in relation to requiring a course of study in the awareness and prevention of human trafficking PURPOSE: To require a course of study-in our education system-concerning the awareness and prevention of human trafficking. SUMMARY OF PROVISIONS: This legislation amends the education law to require a course of study- in our education system-concerning the awareness and prevention of human trafficking.