Bill S339 version original - Robert Jackson
BILL NUMBER: S339 SPONSOR: JACKSON TITLE OF BILL: An act to amend the tax law, in relation to the imposition of a commer- cial vacancy tax PURPOSE: This bill will impose a tax on commercial storefronts that remain vacant for a prolonged period of time. SUMMARY OF PROVISIONS: Section 1 amends the Tax Law by adding a new Article 30-C that creates a commercial vacancy tax. Section 1350, paragraph 1 provides the definition of "commercial store- front.