Bill S345 version original - Andrew Gounardes
BILL NUMBER: S345 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the civil practice law and rules, in relation to the statute of limitations for actions based upon unlawful discriminatory practice in employment PURPOSE OR GENERAL IDEA OF BILL: This bill would extend the statute of limitations for discrimination in the workplace to six years. SUMMARY OF PROVISIONS: Section one amends section 213 of the civil practice law and rules. Section two sets forth the effective date.