Bill S407 version original - Patrick M. Gallivan
BILL NUMBER: S407 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the executive law and the penal law, in relation to clemency PURPOSE OR GENERAL IDEA OF BILL: To strip the governor of the power to grant reprieves, commutations, and pardons, after conviction, for certain offenses. SUMMARY OF PROVISIONS: Section 1: Amends Section 15 of the executive law. Section 2: Amends Subdivision 5 of section 70.00 of the penal law. Section 3: Provides the effective date.