Bill S374 version original - Kevin S. Parker
BILL NUMBER: S374 SPONSOR: PARKER TITLE OF BILL: An act to amend the public service law, in relation to approval of elec- tric transmission line facilities PURPOSE: The purpose to this bill is to require the Public Service Commission (PSC), to consult the most recent State Energy Plan (SEP) before issuing a certificate of environmental compatibility and public need ("certif- icate") for an electric transmission line. SUMMARY OF PROVISIONS: This bill amends paragraph (e) of subsection 1.