Bill S385A version original - George Borrello
BILL NUMBER: S385 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the criminal procedure law, in relation to peace officer status of public safety officers of the Chautauqua Institution within the grounds of and properties owned by the Chautauqua Institution PURPOSE: To designate, as peace officers, public safety officers of the Chautau- qua Institution within the grounds of and properties owned by the Chau- tauqua Institution. SUMMARY OF PROVISIONS: Section 1. Amends Section 2.