Bill S408 version original - Patrick M. Gallivan
BILL NUMBER: S408 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the penal law, in relation to making life imprisonment without parole mandatory for defendants convicted of murder in the first degree or second degree and the victim is a police officer, peace offi- cer, first responder or correctional officer PURPOSE OR GENERAL IDEA OF BILL: To ensure justice is served in cases where a first responder is murdered. SUMMARY OF PROVISIONS: Section 1: Section 60.