Bill S430 version original - Kevin Thomas
BILL NUMBER: S430 SPONSOR: THOMAS TITLE OF BILL: An act to amend the penal law, in relation to prohibiting the possession and use of live ammunition at a film production facility or a qualified film production facility PURPOSE: To prevent studio accidents by prohibiting live ammunition on studio sets and requiring training for the safe handling, use, and storage of firearms. SUMMARY OF PROVISIONS: Section 1 of the bill amends section 265.