Bill S414 version original - Kevin Thomas
BILL NUMBER: S414 SPONSOR: THOMAS TITLE OF BILL: An act to amend the general business law, in relation to prohibiting the sale of sodium nitrate PURPOSE GENERAL IDEA OF BILL: To prevent the sale and shipment of a toxic compound in New York by any individual or entity. SUMMARY OF PROVISIONS: Section one of the bill amends the general business law by adding a new section 398-g to prohibit the sale of sodium nitrate to any person in New York.