Bill S413 version original - Kevin Thomas
BILL NUMBER: S413 SPONSOR: THOMAS TITLE OF BILL: An act to establish the people with disabilities access to programs commission PURPOSE: To streamline the process for people with disabilities in New York to obtain services. SUMMARY OF PROVISIONS: Section 1 establishes a People With Disabilities Access to Programs Commission that is directed to make recommendations for new laws to streamline the process for people with disabilities to access programs and services.