Bill S365B version original - Kevin Thomas
BILL NUMBER: S365 SPONSOR: THOMAS TITLE OF BILL: An act to amend the general business law, in relation to the management and oversight of personal data PURPOSE GENERAL IDEA OF BILL: The purpose of the bill is to help New Yorkers regain their privacy. The bill, cited as the NY Privacy Act, will require the companies to obtain consent from consumers before processing the personal data of consumers. SUMMARY OF PROVISIONS: Section 1 of the bill defines the act as the "New York Privacy Act".