Bill S396 version original - Cordell Cleare
BILL NUMBER: S396 SPONSOR: CLEARE TITLE OF BILL: An act to amend the executive law, in relation to expediting the distributions of funds for the supervision and treatment services for juveniles program PURPOSE OR GENERAL IDEA OF BILL: To expedite the distributions of funds for the supervision and treatment services for juveniles program. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Section 529-b of the executive law is amended by adding a subdivision 6. Section 2: Sets the effective date.