Bill S521 version original - Leroy Comrie
BILL NUMBER: S521 SPONSOR: COMRIE TITLE OF BILL: An act to repeal subdivision 6 of section 51 of the public authorities law, relating to voting by members of the New York state authorities control board PURPOSE OR GENERAL IDEA OF BILL: Removes changes to the PACB statute enacted by demand of the Governor in 2019. SUMMARY OF PROVISIONS: This bill repeals subdivision 6 of section 51 of the public authorities law, relating to voting by members of the New York state authorities control board.