Bill S541B version original - Brad Hoylman-Sigal
BILL NUMBER: S541 SPONSOR: HOYLMAN TITLE OF BILL: An act to amend the labor law, in relation to enacting the "Empowering People in Rights Enforcement (EMPIRE) Worker Protection Act" PURPOSE: This legislation empowers aggrieved workers to file claims on behalf of the state against an employer for specified violations of State labor law. The Act also allows representative organizations to file claims on behalf of impacted workers in order to protect the public interest.