Bill S516 version original - Kevin Thomas
BILL NUMBER: S516 SPONSOR: THOMAS TITLE OF BILL: An act to amend the general business law, in relation to prohibiting the provision of any quantity of an electronic liquid PURPOSE OR GENERAL IDEA OF BILL: To prohibit the sale of refillable electronic liquids, liquid nicotine or e-liquids. SUMMARY OF PROVISIONS: Section 1 of the bill amends the public health law by adding a new section 2 1399-cc-1.