Bill S528 version original - Kevin Thomas
BILL NUMBER: S528 SPONSOR: THOMAS TITLE OF BILL: An act to amend the environmental conservation law and the civil prac- tice law and rules, in relation to establishing a cause of action for medical monitoring PURPOSE OR GENERAL IDEA OF BILL: To create an independent cause of action for medical monitoring. SUMMARY OF PROVISIONS: Section one amends the environmental conservation law to create an inde- pendent cause of action for medical monitoring.